Code for guessing a card, checking if correct colour
Guessing a card for a game will check if it is the correct colour,
reduce the number of guess left based on outcome of guess.  If it is the
black card it will end the game.  If you guess the last word of your
colour it ends the game and sets you as the winner.